[ADD] Backend Finished

+++ backend/app/models/models.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass
 @dataclass
 class VivianiParameters:
     radio: float
-    subintervalos: float
+    subintervalos: int
